Refactor project structure and update to README
Refactored the directory structure by splitting the 'nn' module into 'network' package with submodules for better organization and understandability. Introduced a new 'Device' enumerator in the 'devices' module to handle PyTorch devices which makes it easier to manage devices within the code. Updated 'FeedForwardNetwork' to be a subclass of 'NeuralNetwork' for coherence. 'FeedForwardNetwork' is now more flexible, allowing the user to load network weights. Additionally, updated the 'README' to reflect the changes and improvements made in the project structure and functionality.

This change is meant to enhance code readability and maintainability. It also provides a more comfortable user experience by allowing greater customization of the 'FeedForwardNetwork' and clear description in 'README'.