Update README.md for clarity and completeness.
Enhanced the introduction section of README.md to be more inviting and informative, providing a clearer understanding of our project's purpose. Included a new section for 'Loss' which was missing from the Table of Contents previously. Additionally, wording was updated throughout file to ensure consistency and accuracy such as changing 'file' to 'package' in 'Activations' section. Lastly, enhanced 'Contribute' section to be more inviting, making our project more welcoming for contributors.