Update PyTorch Playground README for clarity and educational focus
Overhauled main README file of PyTorch Playground project to enhance its focus as an educational tool for beginners and experts in deep learning. Changes include a more appealing project title, an introduction section with a license badge, a detailed overview, and a more structured table of contents. It also includes hands-on modules like an MNIST model demo with visual output, comprehensive setup and usage guide. Embedded a call for learners to contribute and collaborate, helping turn the project into a shared learning resource. These changes enhance the readability and user experience of the project, making it more inviting for the learning community.