Add loss and dataset notebooks
The datasets and loss notebooks were added in this commit. The dataset notebook includes markdowns and examples of various datasets objects. The loss notebook has examples and explanations about the Cross-Entropy loss class. The purpose of these changes was to provide detailed explanations and examples of how the different datasets and loss classes work.