Add intro to neural networks and feed-forward network code
This commit introduces a basic setup for exploring neural networks and feed-forward networks, which is the
basis of many machine learning algorithms. We added 01_neural_networks_intro.ipynb and 02_feed_forward_intro.ipynb notebooks under the src/networks directory to discuss the basic concepts related to these network types. Additionally, rectified bugs in the activation function module and removed redundant code for a cleaner implementation. This commit intends to provide users with a basic understanding of how these networks work and follow along with the concepts interactively.